animation and reduce commit

+++ src/hooks/useTodo.js
@@ -1,5 +1,37 @@
-import { useEffect, useState, useMemo, useRef, useCallback } from "react";
+import { useEffect, useState, useMemo, useRef, useCallback, useReducer } from "react";
 import taskAPI from "../api/tasksAPI";
+import useUi from "./useUi";
+
+const tasksReducer = (state, action) => {
+    switch (action.type) {
+        case 'SET_ALL': {
+            return Array.isArray(action.tasks) ? action.tasks : state
+        }
+        case 'ADD': {
+            return [...state, action.tasks]
+        }
+        case 'TOGGLE_COMPLETE': {
+            const { id, isDone } = action
+
+            return state.map((task) => {
+                return task.id === id ? {...task, isDone} : task
+            })
+        }
+        case 'DELETE': {
+            return state.filter(task => task.id !== action.id)
+        }
+        case 'EDIT_TITLE': {
+            const { id, title } = action;
+
+            return state.map((task) => {
+                return task.id === id ? {...task, title} : task
+            })
+        }
+        default: {
+            return state
+        }
+    }
+}
 
 const filterdOptions = {
         all: 'all',
@@ -8,49 +40,70 @@ const filterdOptions = {
 }
 
 export const useTodo = () => {
-    const [tasks, setTasks] = useState([]);
+    const [tasks, dispatch] = useReducer(tasksReducer, []);
     
-    const [isOpenModalWin, setIsOpenModalWin] = useState(false); 
-    const [openList, setOpenList] = useState(false);
-    const [newTaskTitle, setNewTaskTitle] = useState('');
-    const [searchTaskTitle, setSearchTaskTitle] = useState('');
     const [activeOption, setActiveOption] = useState('all');
+    const [searchTaskTitle, setSearchTaskTitle] = useState('');
+    const [newTaskTitle, setNewTaskTitle] = useState('');
     const [editingTaskTitle, setEditingTaskTitle] = useState('');
     const [editingTaskId, setEditingTaskId] = useState(null);
-    const [errorMessage, setErrorMessage] = useState('');
+    const [isLoading, setIsLoading] = useState(true);
+    const {addAnimation, removeAnimation, animation} = useUi()
+    
 
     const editInputRef = useRef(null);
     
-    const addTasks = useCallback(() => {
-        if (newTaskTitle.trim().length === 0) return;
+    const addTasks = useCallback((title) => {
+        if (title.trim().length === 0) return;
 
+        
         const newTask = {
-                title: newTaskTitle,
-                isDone: false,
+            title,
+            isDone: false,
         }
         
         taskAPI.add(newTask)
         .then((data) => {
-            setTasks(tasks => [...tasks, data]);
-            setNewTaskTitle('');
+            dispatch({ type: 'ADD', tasks: data,});
+            addAnimation('apperingAnim', data.id)
+            setTimeout(() => {
+                removeAnimation('apperingAnim', data.id)
+            }, 400);
+        }).catch((data) => {
+            removeAnimation('apperingAnim', data.id)
         })
-    
-    }, [newTaskTitle])
+    }, [addAnimation, removeAnimation])
 
     useEffect(() => {
         taskAPI.getAll().then((data) => {
-            setTasks(data);
+            dispatch({ type: 'SET_ALL', tasks: data,});
+        }).catch(() => {
+            setIsLoading(false);
+        }).finally(() => {
+            setIsLoading(false)
         })
     }, [])
 
     const deleteTask = useCallback((taskId) => {
+        if (animation.deleatingTasks.includes(taskId)) return;
+        addAnimation('deleatingTasks', taskId);
+
         taskAPI.delete(taskId).then(() => {
-            setTasks(prevTasks => prevTasks.filter((task) => task.id !== taskId))
+            addAnimation('deleteAnim', taskId);
+
+            setTimeout(() => {
+                dispatch({ type: 'DELETE', id: taskId,})
+                removeAnimation('deleatingTasks', taskId)
+                removeAnimation('deleteAnim', taskId)
+            }, 400)
+        }).catch(() => {
+            removeAnimation('deleatingTasks', taskId)
+            removeAnimation('deleteAnim', taskId)
         })
-    }, [])
+    }, [animation.deleatingTasks, addAnimation, removeAnimation])
 
     const startEditTask = useCallback((id) => {
-        tasks.find((task) => {
+        tasks.forEach((task) => {
             if (task.id === id) {
                 setEditingTaskTitle(task.title)
                 setEditingTaskId(task.id)
@@ -76,8 +129,6 @@ export const useTodo = () => {
                 setEditingTaskId(null)
             } else if (code === 'Tab') {
                 setEditingTaskId(null)
-            } else if (code === 'Enter') {
-                // 
             }
         }
 
@@ -90,24 +141,17 @@ export const useTodo = () => {
         }
     }, [editingTaskId])
 
-    const editTaskApply = useCallback(
-        (id) => {
-        const title = editingTaskTitle.trim()
+    const editTaskApply = useCallback((id, newTitle) => {
+        const title = newTitle.trim()
 
         if (title.length === 0) return;
 
         taskAPI.editTask(id, title).then(() => {
-            setTasks(tasks => tasks.map((task) => {
-                if (task.id === id) {
-                    return {...task, title}
-                }
-
-                return task
-            }))
+            dispatch({ type: 'EDIT_TITLE', id, title})
         })
         
         setEditingTaskId(null)
-    }, [editingTaskTitle])
+    }, [])
     
     const filteredTasks = useMemo(() => {
     let result = tasks;
@@ -131,21 +175,12 @@ export const useTodo = () => {
     const toggleCheckedTask = useCallback(
         (id, isDone) => {
         taskAPI.toggleComplete(id, isDone).then(() => {
-            setTasks(tasks => tasks.map((task) => {
-                    if(task.id === id) {
-                        return {...task, isDone}
-                    }
-    
-                    return task
-                }))
+            dispatch({ type: 'TOGGLE_COMPLETE', id, isDone})
         })}, [])
 
     return {
         tasks,
-        isOpenModalWin,
-        setIsOpenModalWin,
         toggleCheckedTask,
-        setTasks,
         newTaskTitle,
         setNewTaskTitle,
         searchTaskTitle,
@@ -160,11 +195,9 @@ export const useTodo = () => {
         setEditingTaskTitle,
         editingTaskId, 
         setEditingTaskId,
-        openList,
-        setOpenList,
         editInputRef,
         editTaskApply,
-        errorMessage, 
-        setErrorMessage,
+        isLoading,
+        animation
     }
 }
